# Import/Perf: create edit bones in batches