Previously we would switch to edit mode, create an edit bone, then switch back to object mode for every single bone, which was slow.

````
for each bone:
    switch to edit mode
    create edit bone
    switch to object mode
    do pose bone stuff
````

This switches to edit mode once for the whole armature, creates all bones, and then switches back.

````
for each arma:
    switch to edit mode
    create all edit bones
    switch to object mode
    do pose bone stuff
````

Performance:

| model | master | this branch |
|:---|:---|:---|
| Ruby Rose (#848) | 3.5s | 1.5s |
| Level (#862) | 11.3s | 8.9s |